refactor: function crawl
