+++ src/constants/moment.ts
@@ -1,0 +1,5 @@
+export enum MOMENT_CONSTANT {
+  FORMAT = 'YYYY-MM-DD',
+  YEARS = 'years',
+  DAYS = 'days'
+}
