+++ src/routes/index.ts
@@ -1,11 +1,11 @@
 import * as express from 'express';
 import ApiRouter from './api';
-import CrawlRouter from './craw';
+import FunctionRouter from './function';
 
 const router = express.Router();
 
 /* interface for Rest full api */
 router.use('/api', ApiRouter);
-router.use('/job', CrawlRouter);
+router.use('/job', FunctionRouter);
 
 export default router;
